+++ src/utilities/index.ts
@@ -1,3 +1,3 @@
-export { createFieldMapper } from './fieldMapper';
+export { createFieldMapper, defaultFieldMapper } from './fieldMapper';
 export { guess } from './guess';
 export { inspect } from './inspect';